Added min DFA to search first tests
