+++ 2/myre/CMakeLists.txt
@@ -4,6 +4,8 @@ add_subdirectory(Regex)
 
 target_include_directories(MyReLib PUBLIC include)
 
-target_sources(MyReLib PRIVATE src/myre.cpp)
+target_sources(MyReLib PRIVATE src/compile.cpp
+							   src/search.cpp
+							   src/fullmatch.cpp)
 
 target_link_libraries(MyReLib PUBLIC RegexLib)
