+++ 2/myre/src/compile.cpp
@@ -1,0 +1,17 @@
+#include "myre.hpp"
+
+namespace myre
+{
+	
+DFA compile(const std::string& regex, bool optimize){
+	RegexParser parser;
+	DFABuilder builder;
+	Context context;
+	std::shared_ptr<SyntaxNode> tree = parser.parse(regex, context);
+	DFA dfa = builder.buildDFA(tree, context);
+	if (optimize){
+		dfa = builder.minimize_dfa(dfa, context);
+	}
+	return dfa;
+}
+}
